+++ platform-tests/src/test/java/org/eclipse/deeplearning4j/dl4jcore/gradientcheck/TestOutputLayerGradientChecks.java
@@ -53,7 +53,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 @Tag(TagNames.TRAINING)
 @Tag(TagNames.DL4J_OLD_API)
 @NativeTag
-public class OutputLayerGradientChecks extends BaseDL4JTest {
+public class TestOutputLayerGradientChecks extends BaseDL4JTest {
 
     private static final boolean PRINT_RESULTS = true;
