+++ platform-tests/src/test/java/org/eclipse/deeplearning4j/dl4jcore/regressiontest/TestRegressionTest050.java
@@ -55,7 +55,7 @@ import java.io.File;
 import static org.junit.jupiter.api.Assertions.*;
 @NativeTag
 @Tag(TagNames.DL4J_OLD_API)
-public class RegressionTest050 extends BaseDL4JTest {
+public class TestRegressionTest050 extends BaseDL4JTest {
 
     @Override
     public long getTimeoutMilliseconds() {
