+++ platform-tests/src/test/java/org/eclipse/deeplearning4j/dl4jcore/regressiontest/TestRegressionTest060.java
@@ -58,7 +58,7 @@ import java.io.File;
 import static org.junit.jupiter.api.Assertions.*;
 @NativeTag
 @Tag(TagNames.DL4J_OLD_API)
-public class RegressionTest060 extends BaseDL4JTest {
+public class TestRegressionTest060 extends BaseDL4JTest {
 
     @Override
     public DataType getDataType(){
