+++ platform-tests/src/test/java/org/eclipse/deeplearning4j/dl4jcore/regressiontest/TestRegressionTest100b3.java
@@ -20,31 +20,27 @@
 
 package org.eclipse.deeplearning4j.dl4jcore.regressiontest;
 
-import lombok.extern.slf4j.Slf4j;
 import org.deeplearning4j.BaseDL4JTest;
 import org.eclipse.deeplearning4j.dl4jcore.TestUtils;
-import org.deeplearning4j.nn.api.Layer;
 import org.deeplearning4j.nn.conf.BackpropType;
 import org.deeplearning4j.nn.conf.ConvolutionMode;
 import org.deeplearning4j.nn.conf.graph.LayerVertex;
-import org.deeplearning4j.nn.conf.layers.BatchNormalization;
 import org.deeplearning4j.nn.conf.layers.ConvolutionLayer;
-import org.deeplearning4j.nn.conf.layers.GravesLSTM;
+import org.deeplearning4j.nn.conf.layers.DenseLayer;
+import org.deeplearning4j.nn.conf.layers.LSTM;
 import org.deeplearning4j.nn.conf.layers.RnnOutputLayer;
 import org.deeplearning4j.nn.conf.layers.variational.VariationalAutoencoder;
 import org.deeplearning4j.nn.graph.ComputationGraph;
 import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
 import org.deeplearning4j.nn.weights.WeightInitXavier;
+import org.eclipse.deeplearning4j.dl4jcore.regressiontest.customlayer100a.CustomLayer;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.nd4j.common.tests.tags.NativeTag;
 import org.nd4j.common.tests.tags.TagNames;
-import org.nd4j.linalg.activations.impl.ActivationIdentity;
-import org.nd4j.linalg.activations.impl.ActivationLReLU;
-import org.nd4j.linalg.activations.impl.ActivationSoftmax;
-import org.nd4j.linalg.activations.impl.ActivationTanH;
+import org.nd4j.linalg.activations.impl.*;
 import org.nd4j.linalg.api.buffer.DataType;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.factory.Nd4j;
@@ -56,14 +52,13 @@ import org.nd4j.common.resources.Resources;
 import java.io.DataInputStream;
 import java.io.File;
 import java.io.FileInputStream;
+import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
-
-@Slf4j
 @Disabled
 @NativeTag
 @Tag(TagNames.DL4J_OLD_API)
-public class RegressionTest100a extends BaseDL4JTest {
+public class TestRegressionTest100b3 extends BaseDL4JTest {
 
     @Override
     public long getTimeoutMilliseconds() {
@@ -77,61 +72,99 @@ public class RegressionTest100a extends BaseDL4JTest {
 
     @Test
     public void testCustomLayer() throws Exception {
-        //We dropped support for 1.0.0-alpha and earlier custom layers due to the maintenance overhead for a rarely used feature
-        //An upgrade path exists as a workaround - load in beta to beta4 and re-save
-        //All built-in layers can be loaded going back to 0.5.0
 
-        File f = Resources.asFile("regression_testing/100a/CustomLayerExample_100a.bin");
+        for( int i=1; i<2; i++ ) {
+            String dtype = (i == 0 ? "float" : "double");
+            DataType dt = (i == 0 ? DataType.FLOAT : DataType.DOUBLE);
 
-        try {
+            File f = Resources.asFile("regression_testing/100b3/CustomLayerExample_100b3_" + dtype + ".bin");
             MultiLayerNetwork.load(f, true);
-            fail("Expected exception");
-        } catch (Exception e){
-            String msg = e.getMessage();
-            assertTrue(msg.contains("custom") && msg.contains("1.0.0-beta") && msg.contains("saved again"), msg);
+
+            MultiLayerNetwork net = MultiLayerNetwork.load(f, true);
+//            net = net.clone();
+
+            DenseLayer l0 = (DenseLayer) net.getLayer(0).conf().getLayer();
+            assertEquals(new ActivationTanH(), l0.getActivationFn());
+            Assertions.assertEquals(new WeightDecay(0.03, false), TestUtils.getWeightDecayReg(l0));
+            assertEquals(new RmsProp(0.95), l0.getIUpdater());
+
+            CustomLayer l1 = (CustomLayer) net.getLayer(1).conf().getLayer();
+            assertEquals(new ActivationTanH(), l1.getActivationFn());
+            assertEquals(new ActivationSigmoid(), l1.getSecondActivationFunction());
+            assertEquals(new RmsProp(0.95), l1.getIUpdater());
+
+
+            INDArray outExp;
+            File f2 = Resources.asFile("regression_testing/100b3/CustomLayerExample_Output_100b3_" + dtype + ".bin");
+            try (DataInputStream dis = new DataInputStream(new FileInputStream(f2))) {
+                outExp = Nd4j.read(dis);
+            }
+
+            INDArray in;
+            File f3 = Resources.asFile("regression_testing/100b3/CustomLayerExample_Input_100b3_" + dtype + ".bin");
+            try (DataInputStream dis = new DataInputStream(new FileInputStream(f3))) {
+                in = Nd4j.read(dis);
+            }
+
+            assertEquals(dt, in.dataType());
+            assertEquals(dt, outExp.dataType());
+            assertEquals(dt, net.params().dataType());
+            assertEquals(dt, net.getFlattenedGradients().dataType());
+            assertEquals(dt, net.getUpdater().getStateViewArray().dataType());
+
+            //System.out.println(Arrays.toString(net.params().data().asFloat()));
+
+            INDArray outAct = net.output(in);
+            assertEquals(dt, outAct.dataType());
+
+            List<INDArray> activations = net.feedForward(in);
+
+            assertEquals(dt, net.getLayerWiseConfigurations().getDataType());
+            assertEquals(dt, net.params().dataType());
+            assertEquals(outExp, outAct, dtype);
         }
     }
 
 
     @Test
-    public void testGravesLSTM() throws Exception {
+    public void testLSTM() throws Exception {
 
-        File f = Resources.asFile("regression_testing/100a/GravesLSTMCharModelingExample_100a.bin");
+        File f = Resources.asFile("regression_testing/100b3/GravesLSTMCharModelingExample_100b3.bin");
         MultiLayerNetwork net = MultiLayerNetwork.load(f, true);
 
-        GravesLSTM l0 = (GravesLSTM) net.getLayer(0).conf().getLayer();
+        LSTM l0 = (LSTM) net.getLayer(0).conf().getLayer();
         assertEquals(new ActivationTanH(), l0.getActivationFn());
         assertEquals(200, l0.getNOut());
         assertEquals(new WeightInitXavier(), l0.getWeightInitFn());
-        Assertions.assertEquals(new WeightDecay(0.001, false), TestUtils.getWeightDecayReg(l0));
-        assertEquals(new RmsProp(0.1), l0.getIUpdater());
+        assertEquals(new WeightDecay(0.0001, false), TestUtils.getWeightDecayReg(l0));
+        assertEquals(new Adam(0.005), l0.getIUpdater());
 
-        GravesLSTM l1 = (GravesLSTM) net.getLayer(1).conf().getLayer();
+        LSTM l1 = (LSTM) net.getLayer(1).conf().getLayer();
         assertEquals(new ActivationTanH(), l1.getActivationFn());
         assertEquals(200, l1.getNOut());
         assertEquals(new WeightInitXavier(), l1.getWeightInitFn());
-        assertEquals(new WeightDecay(0.001, false), TestUtils.getWeightDecayReg(l1));
-        assertEquals(new RmsProp(0.1), l1.getIUpdater());
+        assertEquals(new WeightDecay(0.0001, false), TestUtils.getWeightDecayReg(l1));
+        assertEquals(new Adam(0.005), l1.getIUpdater());
 
         RnnOutputLayer l2 = (RnnOutputLayer) net.getLayer(2).conf().getLayer();
         assertEquals(new ActivationSoftmax(), l2.getActivationFn());
         assertEquals(77, l2.getNOut());
         assertEquals(new WeightInitXavier(), l2.getWeightInitFn());
-        assertEquals(new WeightDecay(0.001, false), TestUtils.getWeightDecayReg(l0));
-        assertEquals(new RmsProp(0.1), l0.getIUpdater());
+        assertEquals(new WeightDecay(0.0001, false), TestUtils.getWeightDecayReg(l0));
+        assertEquals(new Adam(0.005), l0.getIUpdater());
 
         assertEquals(BackpropType.TruncatedBPTT, net.getLayerWiseConfigurations().getBackpropType());
         assertEquals(50, net.getLayerWiseConfigurations().getTbpttBackLength());
         assertEquals(50, net.getLayerWiseConfigurations().getTbpttFwdLength());
 
         INDArray outExp;
-        File f2 = Resources.asFile("regression_testing/100a/GravesLSTMCharModelingExample_Output_100a.bin");
+        File f2 = Resources.asFile("regression_testing/100b3/GravesLSTMCharModelingExample_Output_100b3.bin");
         try(DataInputStream dis = new DataInputStream(new FileInputStream(f2))){
             outExp = Nd4j.read(dis);
         }
 
         INDArray in;
-        File f3 = Resources.asFile("regression_testing/100a/GravesLSTMCharModelingExample_Input_100a.bin");
+        File f3 = Resources.asFile("regression_testing/100b3/GravesLSTMCharModelingExample_Input_100b3.bin");
         try(DataInputStream dis = new DataInputStream(new FileInputStream(f3))){
             in = Nd4j.read(dis);
         }
@@ -144,7 +177,7 @@ public class RegressionTest100a extends BaseDL4JTest {
     @Test
     public void testVae() throws Exception {
 
-        File f = Resources.asFile("regression_testing/100a/VaeMNISTAnomaly_100a.bin");
+        File f = Resources.asFile("regression_testing/100b3/VaeMNISTAnomaly_100b3.bin");
         MultiLayerNetwork net = MultiLayerNetwork.load(f, true);
 
         VariationalAutoencoder l0 = (VariationalAutoencoder) net.getLayer(0).conf().getLayer();
@@ -154,16 +187,16 @@ public class RegressionTest100a extends BaseDL4JTest {
         assertArrayEquals(new int[]{256, 256}, l0.getDecoderLayerSizes());
                 assertEquals(new WeightInitXavier(), l0.getWeightInitFn());
         assertEquals(new WeightDecay(1e-4, false), TestUtils.getWeightDecayReg(l0));
-        assertEquals(new Adam(0.05), l0.getIUpdater());
+        assertEquals(new Adam(1e-3), l0.getIUpdater());
 
         INDArray outExp;
-        File f2 = Resources.asFile("regression_testing/100a/VaeMNISTAnomaly_Output_100a.bin");
+        File f2 = Resources.asFile("regression_testing/100b3/VaeMNISTAnomaly_Output_100b3.bin");
         try(DataInputStream dis = new DataInputStream(new FileInputStream(f2))){
             outExp = Nd4j.read(dis);
         }
 
         INDArray in;
-        File f3 = Resources.asFile("regression_testing/100a/VaeMNISTAnomaly_Input_100a.bin");
+        File f3 = Resources.asFile("regression_testing/100b3/VaeMNISTAnomaly_Input_100b3.bin");
         try(DataInputStream dis = new DataInputStream(new FileInputStream(f3))){
             in = Nd4j.read(dis);
         }
@@ -175,10 +208,9 @@ public class RegressionTest100a extends BaseDL4JTest {
 
 
     @Test
-    @Disabled("AB 2019/05/23 - Failing on linux-x86_64-cuda-9.2 - see issue #7657")
     public void testYoloHouseNumber() throws Exception {
 
-        File f = Resources.asFile("regression_testing/100a/HouseNumberDetection_100a.bin");
+        File f = Resources.asFile("regression_testing/100b3/HouseNumberDetection_100b3.bin");
         ComputationGraph net = ComputationGraph.load(f, true);
 
         int nBoxes = 5;
@@ -193,80 +225,20 @@ public class RegressionTest100a extends BaseDL4JTest {
         assertArrayEquals(new int[]{1,1}, cl.getKernelSize());
 
         INDArray outExp;
-        File f2 = Resources.asFile("regression_testing/100a/HouseNumberDetection_Output_100a.bin");
+        File f2 = Resources.asFile("regression_testing/100b3/HouseNumberDetection_Output_100b3.bin");
         try(DataInputStream dis = new DataInputStream(new FileInputStream(f2))){
             outExp = Nd4j.read(dis);
         }
 
         INDArray in;
-        File f3 = Resources.asFile("regression_testing/100a/HouseNumberDetection_Input_100a.bin");
+        File f3 = Resources.asFile("regression_testing/100b3/HouseNumberDetection_Input_100b3.bin");
         try(DataInputStream dis = new DataInputStream(new FileInputStream(f3))){
             in = Nd4j.read(dis);
         }
 
-        //Minor bug in 1.0.0-beta and earlier: not adding epsilon value to forward pass for batch norm
-        //Which means: the record output doesn't have this. To account for this, we'll manually set eps to 0.0 here
-        //https://github.com/eclipse/deeplearning4j/issues/5836#issuecomment-405526228
-        for(Layer l : net.getLayers()){
-            if(l.conf().getLayer() instanceof BatchNormalization){
-                BatchNormalization bn = (BatchNormalization) l.conf().getLayer();
-                bn.setEps(0.0);
-            }
-        }
+        INDArray outAct = net.outputSingle(in);
 
-        INDArray outAct = net.outputSingle(in).castTo(outExp.dataType());
-
-        boolean eq = outExp.equalsWithEps(outAct, 1e-4);
-        if(!eq){
-            log.info("Expected: {}", outExp);
-            log.info("Actual: {}", outAct);
-        }
-        assertTrue(eq, "Output not equal");
+        boolean eq = outExp.equalsWithEps(outAct.castTo(outExp.dataType()), 1e-3);
+        assertTrue(eq);
     }
-
-
-    @Test
-    @Disabled("Ignoring due to new set input types changes. Loading a network isn't a problem, but we need to set the input types yet.")
-    public void testUpsampling2d() throws Exception {
-
-        File f = Resources.asFile("regression_testing/100a/upsampling/net.bin");
-        MultiLayerNetwork net = MultiLayerNetwork.load(f, true);
-
-        INDArray in;
-        File fIn = Resources.asFile("regression_testing/100a/upsampling/in.bin");
-        try(DataInputStream dis = new DataInputStream(new FileInputStream(fIn))){
-            in = Nd4j.read(dis);
-        }
-
-
-        INDArray label;
-        File fLabels = Resources.asFile("regression_testing/100a/upsampling/labels.bin");
-        try(DataInputStream dis = new DataInputStream(new FileInputStream(fLabels))){
-            label = Nd4j.read(dis);
-        }
-
-        INDArray outExp;
-        File fOutExp = Resources.asFile("regression_testing/100a/upsampling/out.bin");
-        try(DataInputStream dis = new DataInputStream(new FileInputStream(fOutExp))){
-            outExp = Nd4j.read(dis);
-        }
-
-        INDArray gradExp;
-        File fGradExp = Resources.asFile("regression_testing/100a/upsampling/gradient.bin");
-        try(DataInputStream dis = new DataInputStream(new FileInputStream(fGradExp))){
-            gradExp = Nd4j.read(dis);
-        }
-
-        INDArray out = net.output(in, false);
-        assertEquals(outExp, out);
-
-        net.setInput(in);
-        net.setLabels(label);
-        net.computeGradientAndScore();
-
-        INDArray grad = net.getFlattenedGradients();
-        assertEquals(gradExp, grad);
-    }
-
-
 }
