+++ platform-tests/src/test/java/org/eclipse/deeplearning4j/nd4j/autodiff/opvalidation/TestActivationGradChecks.java
@@ -44,7 +44,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 @NativeTag
 @Tag(TagNames.SAMEDIFF)
 @TrainingTag
-public class ActivationGradChecks extends BaseOpValidation {
+public class TestActivationGradChecks extends BaseOpValidation {
 
 
     @ParameterizedTest
