+++ platform-tests/src/test/java/org/eclipse/deeplearning4j/nd4j/autodiff/opvalidation/TestLossOpValidation.java
@@ -44,7 +44,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @Slf4j
 @NativeTag
-public class LossOpValidation extends BaseOpValidation {
+public class TestLossOpValidation extends BaseOpValidation {
 
 
     @Override
