+++ platform-tests/src/test/java/org/eclipse/deeplearning4j/nd4j/autodiff/opvalidation/TestReductionBpOpValidation.java
@@ -50,7 +50,7 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 
 @Slf4j
 @NativeTag
-public class ReductionBpOpValidation extends BaseOpValidation {
+public class TestReductionBpOpValidation extends BaseOpValidation {
 
     private DataType initialType;
