+++ platform-tests/src/test/java/org/eclipse/deeplearning4j/nd4j/autodiff/opvalidation/TestShapeOpValidation.java
@@ -72,7 +72,7 @@ import static org.nd4j.linalg.indexing.NDArrayIndex.*;
 @Slf4j
 @NativeTag
 @Tag(TagNames.SAMEDIFF)
-public class ShapeOpValidation extends BaseOpValidation {
+public class TestShapeOpValidation extends BaseOpValidation {
 
     /*
     To test:
