+++ platform-tests/src/test/java/org/eclipse/deeplearning4j/nd4j/linalg/lapack/TestLapackF.java
@@ -23,13 +23,10 @@ package org.eclipse.deeplearning4j.nd4j.linalg.lapack;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import org.nd4j.common.tests.tags.NativeTag;
-import org.nd4j.common.tests.tags.TagNames;
 import org.nd4j.linalg.BaseNd4jTestWithBackends;
 import org.nd4j.linalg.api.buffer.DataType;
 import org.nd4j.linalg.api.ndarray.INDArray;
@@ -40,7 +37,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @Slf4j
 @NativeTag
-public class LapackTestsC extends BaseNd4jTestWithBackends {
+public class TestLapackF extends BaseNd4jTestWithBackends {
     DataType initialType = Nd4j.dataType();
 
     @BeforeEach
@@ -53,12 +50,12 @@ public class LapackTestsC extends BaseNd4jTestWithBackends {
         Nd4j.setDataType(initialType);
     }
 
-    @ParameterizedTest
+      @ParameterizedTest
     @MethodSource("org.nd4j.linalg.BaseNd4jTestWithBackends#configs")
     public void testGetRF1DifferentOrders(Nd4jBackend backend) {
-        INDArray a = Nd4j.linspace(1, 9, 9, Nd4j.dataType()).reshape(3, 3);
+        INDArray a = Nd4j.create(new double[] {1, 2, 3, 4, 5, 6, 7, 8, 9}, new int[] {3, 3}, 'c').dup('f');
         INDArray exp = Nd4j.create(new double[] {7.0, 8.0, 9.0, 0.14285715, 0.85714287, 1.7142857, 0.5714286, 0.5, 0.0},
-                new int[] {3, 3}, 'c');
+                        new int[] {3, 3}, 'c').dup('f');
 
         INDArray r = Nd4j.getNDArrayFactory().lapack().getrf(a);
 
@@ -67,6 +64,6 @@ public class LapackTestsC extends BaseNd4jTestWithBackends {
 
     @Override
     public char ordering() {
-        return 'c';
+        return 'f';
     }
 }
